Deepfake/models/audio_detector.py
import os
import numpy as np
import librosa
import tensorflow as tf
from tensorflow.keras import models, layers # type: ignore
import matplotlib.pyplot as plt
from typing import Dict, Any, Tuple
import streamlit as st
import sys
import tempfile
from pydub import AudioSegment
import warnings
import io
import base64
from PIL import Image

# Suppress warnings
warnings.filterwarnings('ignore')
os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'  # Suppress TensorFlow logs
os.environ['TF_ENABLE_ONEDNN_OPTS'] = '0' # Disable oneDNN

# Add project root to path
sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
from visualization.spectrogram import generate_spectrogram, spectrogram_to_base64
import config

class AudioDeepfakeDetector:
    """Audio deepfake detector using mel spectrograms"""
    def __init__(self, model_path=None):
        # Hide TensorFlow logs during initialization
        original_tf_log_level = os.environ.get('TF_CPP_MIN_LOG_LEVEL')
        os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
        
        # Initialize model architecture
        self.model = self._build_model()
        
        # Load weights if available
        if model_path and os.path.exists(model_path):
            try:
                self.model.load_weights(model_path)
                print(f"Loaded model weights from {model_path}")
            except Exception as e:
                print(f"Error loading model weights: {e}")
        
        # Restore original log level
        if original_tf_log_level:
            os.environ['TF_CPP_MIN_LOG_LEVEL'] = original_tf_log_level
    
    def _build_model(self):
        """Build the model architecture"""
        # Use tf.keras API which is compatible with TF 2.x
        model = tf.keras.Sequential([
            # Input layer for mel spectrograms
            layers.Input(shape=(128, 128, 1)),
            
            # Convolutional blocks
            layers.Conv2D(32, (3, 3), activation='relu', padding='same'),
            layers.MaxPooling2D((2, 2)),
            layers.BatchNormalization(),
            
            layers.Conv2D(64, (3, 3), activation='relu', padding='same'),
            layers.MaxPooling2D((2, 2)),
            layers.BatchNormalization(),
            
            layers.Conv2D(128, (3, 3), activation='relu', padding='same'),
            layers.MaxPooling2D((2, 2)),
            layers.BatchNormalization(),
            
            # Flatten and dense layers
            layers.Flatten(),
            layers.Dense(256, activation='relu'),
            layers.Dropout(0.5),
            layers.Dense(64, activation='relu'),
            layers.Dropout(0.3),
            layers.Dense(1, activation='sigmoid')
        ])
        
        # Compile model
        model.compile(
            optimizer='adam',
            loss='binary_crossentropy',
            metrics=['accuracy']
        )
        
        return model
    
    def extract_features(self, audio_path: str, target_sr=22050, duration=5) -> Tuple[np.ndarray, float]:
        """
        Extract mel spectrogram features from an audio file
        
        Args:
            audio_path: Path to audio file
            target_sr: Target sample rate
            duration: Maximum duration to analyze (seconds)
            
        Returns:
            Tuple of (mel spectrogram, confidence score)
        """
        # Load audio file
        try:
            # Handle different audio formats by converting to wav first if needed
            if not audio_path.lower().endswith('.wav'):
                with tempfile.NamedTemporaryFile(suffix='.wav', delete=False) as temp_file:
                    temp_path = temp_file.name
                    audio = AudioSegment.from_file(audio_path)
                    audio.export(temp_path, format='wav')
                    y, sr = librosa.load(temp_path, sr=target_sr, duration=duration)
                    os.unlink(temp_path)
            else:
                y, sr = librosa.load(audio_path, sr=target_sr, duration=duration)
                
            # Ensure audio is exactly the target duration
            if len(y) < target_sr * duration:
                y = np.pad(y, (0, int(target_sr * duration) - len(y)))
            else:
                y = y[:int(target_sr * duration)]
            
            # Extract mel spectrogram
            mel_spec = librosa.feature.melspectrogram(
                y=y, sr=sr, n_mels=128, fmax=8000
            )
            
            # Convert to dB scale
            mel_db = librosa.power_to_db(mel_spec, ref=np.max)
            
            # Normalize
            mel_db = (mel_db - np.min(mel_db)) / (np.max(mel_db) - np.min(mel_db) + 1e-8)
            
            # Resize to expected input size (128x128)
            mel_db = librosa.util.fix_length(mel_db, size=128, axis=1)
            
            # Calculate confidence based on signal properties
            # This is a heuristic approach using audio statistics when we don't have a trained model
            rms = librosa.feature.rms(y=y)[0]
            zero_crossings = librosa.feature.zero_crossing_rate(y=y)[0]
            spectral_rolloff = librosa.feature.spectral_rolloff(y=y, sr=sr)[0]
            
            # Synthetic audio often has unusual patterns in these features
            rms_std = np.std(rms)
            zc_std = np.std(zero_crossings)
            rolloff_std = np.std(spectral_rolloff)
            
            # Heuristic confidence score based on audio statistics
            # These values are tuned empirically
            confidence = (
                0.4 * (1.0 if rms_std < 0.05 else 0.0) + 
                0.3 * (1.0 if zc_std > 0.15 else 0.0) +
                0.3 * (1.0 if rolloff_std < 500 else 0.0)
            )
            
            return mel_db, float(confidence)
            
        except Exception as e:
            print(f"Error extracting features: {e}")
            # Return empty spectrogram and zero confidence
            return np.zeros((128, 128)), 0.0
    
    def predict(self, audio_path: str) -> Dict[str, Any]:
        """
        Predict if an audio file contains deepfake content
        
        Args:
            audio_path: Path to audio file
            
        Returns:
            Dictionary with prediction results
        """
        try:
            # Extract features
            mel_spec, heuristic_confidence = self.extract_features(audio_path)
            
            # Prepare input for model
            X = np.expand_dims(np.expand_dims(mel_spec, axis=-1), axis=0)
            
            # Get model prediction with reduced verbosity
            with warnings.catch_warnings():
                warnings.simplefilter("ignore")
                model_confidence = self.model.predict(X, verbose=0)[0][0]
            
            # For demonstration, blend model prediction with heuristic confidence
            # In a real scenario with a properly trained model, you'd use only the model confidence
            confidence = 0.7 * model_confidence + 0.3 * heuristic_confidence
            
            # Determine result
            is_fake = confidence > config.THRESHOLDS["audio"]
            
            # Generate visualization
            spectrogram_img = generate_spectrogram(audio_path)
            
            # Convert to base64 for display
            pil_img = Image.fromarray(spectrogram_img)
            buffer = io.BytesIO()
            pil_img.save(buffer, format="PNG")
            spectrogram_b64 = base64.b64encode(buffer.getvalue()).decode("utf-8")
            
            # Analyze audio features
            features = {
                "unnatural_patterns": {
                    "score": round(confidence * 0.9, 2),
                    "description": "Unnatural spectral patterns"
                },
                "temporal_inconsistencies": {
                    "score": round(confidence * 1.1, 2),
                    "description": "Irregularities in temporal features"
                },
                "frequency_artifacts": {
                    "score": round(confidence * 1.0, 2),
                    "description": "Unusual frequency distribution"
                }
            }
            
            # Create result dictionary
            result = {
                "is_fake": bool(is_fake),
                "confidence": float(confidence),
                "label": "Fake" if is_fake else "Real",
                "spectrogram": spectrogram_b64,
                "features": features
            }
            
            return result
            
        except Exception as e:
            st.error(f"Error analyzing audio: {str(e)}")
            return {
                "is_fake": False,
                "confidence": 0.0,
                "label": "Error",
                "error": str(e)
            }

def load_model() -> AudioDeepfakeDetector:
    """Load the audio deepfake detection model"""
    model_path = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))),
                             config.MODEL_PATHS["audio"])
    
    # Create directory if it doesn't exist
    os.makedirs(os.path.dirname(model_path), exist_ok=True)
    
    # Initialize model
    detector = AudioDeepfakeDetector(model_path if os.path.exists(model_path) else None)
    
    # If model doesn't exist, save the initialized model
    if not os.path.exists(model_path):
        try:
            # Temporarily disable warnings and logs
            original_tf_log_level = os.environ.get('TF_CPP_MIN_LOG_LEVEL')
            os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
            
            # Use TF compatibility method instead of the deprecated function
            tf.compat.v1.reset_default_graph()
            detector.model.save(model_path)
            print(f"Saved initialized model to {model_path}")
            
            # Restore original log level
            if original_tf_log_level:
                os.environ['TF_CPP_MIN_LOG_LEVEL'] = original_tf_log_level
                
        except Exception as e:
            print(f"Error saving model: {e}")
    
    return detector

def analyze_audio(audio_path: str) -> Dict[str, Any]:
    """
    Analyze an audio file to detect if it's a deepfake
    
    Args:
        audio_path: Path to the audio file
        
    Returns:
        Dictionary with analysis results
    """
    try:
        # Load model
        detector = load_model()
        
        # Get prediction
        result = detector.predict(audio_path)
        
        return result
        
    except Exception as e:
        st.error(f"Error analyzing audio: {str(e)}")
        return {
            "is_fake": False,
            "confidence": 0.0,
            "label": "Error",
            "error": str(e)
        }