Deepfake/requirements.txt
streamlit==1.30.0
supabase==2.3.0
pillow==10.2.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
librosa==0.10.1
plotly==5.18.0
tensorflow-cpu==2.15.0
scipy==1.12.0
moviepy==1.0.3
python-dotenv==1.0.1
pydub==0.25.1
matplotlib==3.8.3
torch==2.2.1
torchvision==0.17.1
facenet-pytorch==2.5.3